requirements.txt
# Agentic Access Strategy Platform Dependencies

# Web Framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
websockets>=12.0
httpx>=0.27.0

# Data Validation
pydantic>=2.6.0
pydantic-settings>=2.2.0

# Environment
python-dotenv>=1.0.1

# Database
sqlalchemy[asyncio]>=2.0.29
aiosqlite>=0.20.0

# LLM Clients
anthropic>=0.25.0
google-generativeai>=0.5.0
openai>=1.14.0

# Orchestration
langgraph>=0.0.40

# Logging
structlog>=24.1.0

# Utilities
tenacity>=8.2.3
